+++ service-worker.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "pimtc-v16-0-0";
+const CACHE_NAME = "pimtc-v16-0-1";
 const APP_SHELL = [
   "./",
   "./index.html",